\name{Rogers_Adolescent}
\alias{Rogers_Adolescent}
\docType{data}
\title{
Co-Morbid Obsessive-Compulsive Disorder and Depression -- Adolescents
}
\description{
Depression/OCD Data Collected at Rogers Memorial Hospital. The scales used in this study were the Quick Inventory of Depressive Symptomatology self-report version (QIDS-SR), and the Yale-Brown Obsessive Compulsive Scale - self-report (Y-BOCS-SR). The depression scale has 16 items (5 response categories), the OCD scale 10 items (4 response categories).
}

\usage{
data("Rogers_Adolescent")
}

\format{A data frame with 87 individuals and 26 variables (16 depression items followed by 10 OCD items). See \code{?Rogers} for details on individual items.
}

\source{
Jones, P. J., Mair, P., Riemann, B. C., Mugno, B. L., & McNally, R. J. (2018). A network perspective on comorbid depression in adolescents with obsessive-compulsive disorder. Journal of Anxiety Disorders, 53, 1-8.
#'
}


\examples{
data("Rogers_Adolescent")
str(Rogers_Adolescent)
}
\keyword{datasets}
