\name{Wilmer}
\alias{Wilmer}
\docType{data}
\title{
Verbal Paired-Associates Memory Test (VPMT)
}
\description{
Contains data from testmybrain.org within the context of face recognition. It includes the VPMT subscale.  
}

\usage{
data("Wilmer")
}

\format{
A data frame with 1471 individuals, 25 VPMT items, as well as age and gender of the participant. 
}

\source{
Wilmer, J. B., Germine, L., Chabris, C. F., Chatterjee, G., Gerbasi, M. & Nakayama, K. (2012): Capturing specific abilities as a window into human individuality: The example of face recognition, Cognitive Neuropsychology, 29, 360-392
}


\examples{
data(Wilmer)
str(Wilmer)
}
\keyword{datasets}
