\name{WilPat}
\alias{WilPat}
\docType{data}
\title{
Wilson-Patterson Conservatism Scale
}
\description{
This dataset contains a modified version of the classical Wilson-Patterson conservative scale. Each item has the following response categories: 0 ... disapprove, 1 ... approve, 2 ... don't know. 
}
\usage{
data("WilPat")
}
\format{
The first 15 items are conservative items, the remaining ones are liberal. There are 804 persons in the sample. 
}

\note{
Thanks to Benedek Kurdi and Levente Littvay for providing this dataset. 
}

\examples{
data("WilPat")
str(WilPat)
}
\keyword{datasets}
