\name{yaass}
\alias{yaass}
\docType{data}
\title{
YAASS dataset
}
\description{
Contains 30 participants of which 17 are of high risk psychosis and 13 are healthy controls. We have three 
variables pertaining to behavioral measures (factor scores): affective empathy (AE), positive social experience (PSE), and 
perspective taking (PT). Two additional measures come from fMRI scans (right hand fRH and left/right foot fLRF). 
}
\usage{
data("yaass")
}
\format{
 A data frame with 30 observations and 6 variables.
}


\examples{
data("yaass")
str(yaass)
}
\keyword{datasets}
